+++ all/src/test/java/veslo/UniversalCallAdapterFactoryFuncTests.java
@@ -23,24 +23,22 @@ import lombok.experimental.Accessors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import veslo.bean.template.TemplateReplaceAll;
+import veslo.bean.template.TemplateSource;
 import veslo.client.DualResponseAdapterFactoryClient;
 import veslo.client.LoggedMockInterceptor;
 import veslo.client.adapter.UniversalCallAdapterFactory;
 import veslo.client.converter.ExtensionConverterFactory;
 import veslo.client.model.RawBody;
-import veslo.client.model.ReplaceAll;
-import veslo.client.model.ReplaceableTextFile;
 import veslo.client.model.ResourceFile;
 import veslo.util.Utils;
 
 import java.io.File;
-import java.lang.reflect.Field;
 
 import static internal.test.utils.client.MockInterceptor.*;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.hamcrest.Matchers.is;
+import static veslo.bean.template.TemplateSourceType.RESOURCE;
 import static veslo.client.converter.api.ExtensionConverter.NULL_BODY_VALUE;
-import static veslo.client.model.ReplaceableTextFile.TextFileType.RESOURCE_FILE;
 
 @SuppressWarnings("unused")
 @DisplayName("JavaTypeAdapterFactory functional tests")
@@ -274,12 +272,11 @@ public class UniversalCallAdapterFactoryFuncTests extends BaseFuncTest {
                     "        <status>DRAFT</status>\n" +
                     "    </body>\n" +
                     "</note>";
-            CLIENT.returnString(OK, notes.toString()).assertResponse(asserter -> asserter
+            CLIENT.returnString(OK, notes).assertResponse(asserter -> asserter
                     .assertHttpStatusCodeIs(OK)
                     .assertSucBody(BaseUnitTest::assertIs, expected)
                     .assertErrBodyIsNull());
         }
-
 
         @Test
         @DisplayName("String: return String if unsuccessful status and body = string")
@@ -709,22 +706,15 @@ public class UniversalCallAdapterFactoryFuncTests extends BaseFuncTest {
     @Getter
     @Setter
     @Accessors(chain = true, fluent = true)
-    public static final class Notes extends ReplaceableTextFile {
+    @TemplateSource(type = RESOURCE, path = "Notes.xml")
+    public static final class Notes {
 
-        @ReplaceAll(regex = "\\[note.to]")
+        @TemplateReplaceAll(regex = "\\[note.to]")
         private String to;
-        @ReplaceAll(regex = "replace_note_from")
+        @TemplateReplaceAll(regex = "replace_note_from")
         private String from;
-        @ReplaceAll(regex = "\\{note_body_content}")
+        @TemplateReplaceAll(regex = "\\{note_body_content}")
         private String bodyContent;
-
-        public Notes() {
-            super(RESOURCE_FILE, "Notes.xml", UTF_8);
-        }
-
-        public String getReplacement(final Field field) {
-            return super.getReplacement(field);
-        }
 
     }