+++ core/src/main/java/veslo/bean/template/TemplateSource.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2021-2022 Shaburov Oleg
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package veslo.bean.template;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Documented
+@Target(TYPE)
+@Retention(RUNTIME)
+public @interface TemplateSource {
+
+    /**
+     * @return source type ({@link TemplateSourceType#FILE} or {@link TemplateSourceType#RESOURCE})
+     */
+    TemplateSourceType type();
+
+    /**
+     * @return source file path
+     */
+    String path();
+
+    /**
+     * @return source file charset (UTF-8 by default)
+     */
+    String charset() default "UTF-8";
+
+}
+
+