+++ core/src/main/java/veslo/client/model/ResourceFile.java
@@ -16,12 +16,8 @@
 
 package veslo.client.model;
 
-import veslo.ResourceFileException;
 import veslo.util.Utils;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
 import java.nio.charset.Charset;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
@@ -42,52 +38,20 @@ public class ResourceFile {
     public ResourceFile(String resourceRelativePath, Charset charset) {
         Utils.parameterRequireNonNull(resourceRelativePath, "resourceRelativePath");
         Utils.parameterRequireNonNull(charset, "charset");
-        requireExists(resourceRelativePath);
         this.resourceRelativePath = resourceRelativePath;
         this.charset = charset;
     }
 
-    @SuppressWarnings("ResultOfMethodCallIgnored")
-    public byte[] getBytes() {
-        final String path = getResourceRelativePath();
-        try (final InputStream stream = getClassLoader().getResourceAsStream(path)) {
-            if (stream == null) {
-                throw new ResourceFileException("Resource not exists: " + path);
-            }
-            byte[] result = new byte[stream.available()];
-            stream.read(result);
-            return result;
-        } catch (IOException ioException) {
-            throw new ResourceFileException("Resource not readable: " + path, ioException);
-        }
-    }
-
-    @Override
-    public String toString() {
-        return new String(getBytes(), charset);
-    }
-
-    public static String resourceToString(String resourceRelativePath, Charset charset) {
-        return new ResourceFile(resourceRelativePath, charset).toString();
-    }
-
-    public static String resourceToString(String resourceRelativePath) {
-        return new ResourceFile(resourceRelativePath).toString();
+    public String read() {
+        return Utils.readResourceFile(getResourceRelativePath(), getCharset());
     }
 
     public String getResourceRelativePath() {
         return resourceRelativePath;
     }
 
-    protected ClassLoader getClassLoader() {
-        return Thread.currentThread().getContextClassLoader();
-    }
-
-    protected void requireExists(String path) {
-        final URL resource = getClassLoader().getResource(path);
-        if (resource == null) {
-            throw new ResourceFileException("Resource not exists: " + path);
-        }
+    public Charset getCharset() {
+        return charset;
     }
 
 }